requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.23
langchain-google-genai>=0.0.6
langgraph>=0.0.20

# Ollama for offline mode
ollama>=0.1.0

# Search and retrieval
rank-bm25>=0.2.2
tavily-python>=0.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

